chore: Renamed `<CompareYourCostsTrust>` to `<CompareYourTrust>`
