+++ front-end-components/index.html
@@ -88,11 +88,7 @@
       ></div>
       <button type="submit">Submit</button>
     </form>-->
-    <div
-      id="compare-your-costs-trust"
-      data-type="trust"
-      data-id="07465701"
-    ></div>
+    <div id="compare-your-costs-trust" data-id="07465701"></div>
     <!--<div id="compare-your-costs" data-type="school" data-id="140565"></div>-->
     <script type="module" src="/src/main.tsx"></script>
     <script
