--- front-end-components/src/views/compare-your-costs-trust/index.tsx
@@ -1,3 +1,0 @@
-/* eslint-disable react-refresh/only-export-components */
-export * from "src/views/compare-your-costs-trust/types";
-export * from "src/views/compare-your-costs-trust/view";
