--- front-end-components/src/views/compare-your-costs-trust/partials/accordion-sections/index.tsx
@@ -1,10 +1,0 @@
-/* eslint-disable react-refresh/only-export-components */
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/utilities";
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/teaching-support-staff";
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/other-costs";
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/premises-staff-services";
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/non-educational-support-staff";
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/educational-supplies";
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/educational-ict";
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/catering-staff-services";
-export * from "src/views/compare-your-costs-trust/partials/accordion-sections/administrative-supplies";
