--- front-end-components/src/views/compare-your-costs-trust/partials/index.tsx
@@ -1,4 +1,0 @@
-/* eslint-disable react-refresh/only-export-components */
-export * from "src/views/compare-your-costs-trust/partials/total-expenditure";
-export * from "src/views/compare-your-costs-trust/partials/expenditure-accordion";
-export * from "src/views/compare-your-costs-trust/partials/types";
