--- front-end-components/src/views/compare-your-costs-trust/types.tsx
@@ -1,4 +1,0 @@
-export type CompareYourCostsTrustViewProps = {
-  type: string;
-  id: string;
-};
