+++ front-end-components/src/views/compare-your-trust/index.tsx
@@ -1,0 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
+export * from "src/views/compare-your-trust/types";
+export * from "src/views/compare-your-trust/view";
