+++ front-end-components/src/views/compare-your-trust/partials/accordion-sections/administrative-supplies.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { AdministrativeSuppliesData } from "src/views/compare-your-costs-trust/partials/accordion-sections/types";
+import { AdministrativeSuppliesData } from "src/views/compare-your-trust/partials/accordion-sections/types";
 import {
   CostCategories,
   PoundsPerPupil,
@@ -15,21 +15,19 @@ import classNames from "classnames";
 import { TrustExpenditure, ExpenditureApi } from "src/services";
 
 export const AdministrativeSupplies: React.FC<{
-  type: string;
   id: string;
-}> = ({ type, id }) => {
+}> = ({ id }) => {
   const [dimension, setDimension] = useState(PoundsPerPupil);
   const [data, setData] = useState<TrustExpenditure[] | null>();
   const getData = useCallback(async () => {
     setData(null);
     return await ExpenditureApi.trust(
-      type,
       id,
       dimension.value,
       "AdministrationSupplies",
       true
     );
-  }, [id, dimension, type]);
+  }, [id, dimension]);
 
   useEffect(() => {
     getData().then((result) => {
