+++ front-end-components/src/views/compare-your-trust/partials/accordion-sections/teaching-support-staff.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { NonEducationalSupportStaffData } from "src/views/compare-your-costs-trust/partials/accordion-sections/types";
+import { TeachingSupportStaffData } from "src/views/compare-your-trust/partials/accordion-sections/types";
 import {
   CostCategories,
   PoundsPerPupil,
@@ -10,26 +10,22 @@ import {
   HorizontalBarChartWrapper,
   HorizontalBarChartWrapperData,
 } from "src/composed/horizontal-bar-chart-wrapper";
-import classNames from "classnames";
 import { useHash } from "src/hooks/useHash";
+import classNames from "classnames";
 import { TrustExpenditure, ExpenditureApi } from "src/services";
 
-export const NonEducationalSupportStaff: React.FC<{
-  type: string;
-  id: string;
-}> = ({ type, id }) => {
+export const TeachingSupportStaff: React.FC<{ id: string }> = ({ id }) => {
   const [dimension, setDimension] = useState(PoundsPerPupil);
   const [data, setData] = useState<TrustExpenditure[] | null>();
   const getData = useCallback(async () => {
     setData(null);
     return await ExpenditureApi.trust(
-      type,
       id,
       dimension.value,
-      "NonEducationalSupportStaff",
+      "TeachingTeachingSupportStaff",
       true
     );
-  }, [id, dimension, type]);
+  }, [id, dimension]);
 
   useEffect(() => {
     getData().then((result) => {
@@ -56,91 +52,103 @@ export const NonEducationalSupportStaff: React.FC<{
     setDimension(dimension);
   };
 
-  const administrativeClericalBarData: HorizontalBarChartWrapperData<NonEducationalSupportStaffData> =
+  const totalTeachingBarData: HorizontalBarChartWrapperData<TeachingSupportStaffData> =
     useMemo(() => {
       return {
         dataPoints:
           data?.map((trust) => {
             return {
               ...trust,
-              totalValue: trust.administrativeClericalStaffCosts ?? 0,
-              schoolValue: trust.schoolAdministrativeClericalStaffCosts ?? 0,
-              centralValue: trust.centralAdministrativeClericalStaffCosts ?? 0,
+              totalValue: trust.totalTeachingSupportStaffCosts ?? 0,
+              schoolValue: trust.schoolTotalTeachingSupportStaffCosts ?? 0,
+              centralValue: trust.centralTotalTeachingSupportStaffCosts ?? 0,
             };
           }) ?? [],
         tableHeadings,
       };
     }, [data, tableHeadings]);
 
-  const totalNonEducationalBarData: HorizontalBarChartWrapperData<NonEducationalSupportStaffData> =
+  const teachingStaffBarData: HorizontalBarChartWrapperData<TeachingSupportStaffData> =
     useMemo(() => {
       return {
         dataPoints:
           data?.map((trust) => {
             return {
               ...trust,
-              totalValue: trust.totalNonEducationalSupportStaffCosts ?? 0,
-              schoolValue:
-                trust.schoolTotalNonEducationalSupportStaffCosts ?? 0,
-              centralValue:
-                trust.centralTotalNonEducationalSupportStaffCosts ?? 0,
+              totalValue: trust.teachingStaffCosts ?? 0,
+              schoolValue: trust.schoolTeachingStaffCosts ?? 0,
+              centralValue: trust.centralTeachingStaffCosts ?? 0,
             };
           }) ?? [],
         tableHeadings,
       };
     }, [data, tableHeadings]);
 
-  const auditorsCostsBarData: HorizontalBarChartWrapperData<NonEducationalSupportStaffData> =
+  const supplyTeachingBarData: HorizontalBarChartWrapperData<TeachingSupportStaffData> =
     useMemo(() => {
       return {
         dataPoints:
           data?.map((trust) => {
             return {
               ...trust,
-              totalValue: trust.auditorsCosts ?? 0,
-              schoolValue: trust.schoolAuditorsCosts ?? 0,
-              centralValue: trust.centralAuditorsCosts ?? 0,
+              totalValue: trust.supplyTeachingStaffCosts ?? 0,
+              schoolValue: trust.schoolSupplyTeachingStaffCosts ?? 0,
+              centralValue: trust.centralSupplyTeachingStaffCosts ?? 0,
             };
           }) ?? [],
         tableHeadings,
       };
     }, [data, tableHeadings]);
 
-  const otherStaffCostsBarData: HorizontalBarChartWrapperData<NonEducationalSupportStaffData> =
+  const educationalConsultancyBarData: HorizontalBarChartWrapperData<TeachingSupportStaffData> =
     useMemo(() => {
       return {
         dataPoints:
           data?.map((trust) => {
             return {
               ...trust,
-              totalValue: trust.otherStaffCosts ?? 0,
-              schoolValue: trust.schoolOtherStaffCosts ?? 0,
-              centralValue: trust.centralOtherStaffCosts ?? 0,
+              totalValue: trust.educationalConsultancyCosts ?? 0,
+              schoolValue: trust.schoolEducationalConsultancyCosts ?? 0,
+              centralValue: trust.centralEducationalConsultancyCosts ?? 0,
             };
           }) ?? [],
         tableHeadings,
       };
     }, [data, tableHeadings]);
 
-  const professionalServicesBarData: HorizontalBarChartWrapperData<NonEducationalSupportStaffData> =
+  const educationSupportStaffBarData: HorizontalBarChartWrapperData<TeachingSupportStaffData> =
     useMemo(() => {
       return {
         dataPoints:
           data?.map((trust) => {
             return {
               ...trust,
-              totalValue: trust.professionalServicesNonCurriculumCosts ?? 0,
-              schoolValue:
-                trust.schoolProfessionalServicesNonCurriculumCosts ?? 0,
-              centralValue:
-                trust.centralProfessionalServicesNonCurriculumCosts ?? 0,
+              totalValue: trust.educationSupportStaffCosts ?? 0,
+              schoolValue: trust.schoolEducationSupportStaffCosts ?? 0,
+              centralValue: trust.centralEducationSupportStaffCosts ?? 0,
             };
           }) ?? [],
         tableHeadings,
       };
     }, [data, tableHeadings]);
 
-  const elementId = "non-educational-support-staff";
+  const agencySupplyBarData: HorizontalBarChartWrapperData<TeachingSupportStaffData> =
+    useMemo(() => {
+      return {
+        dataPoints:
+          data?.map((trust) => {
+            return {
+              ...trust,
+              totalValue: trust.agencySupplyTeachingStaffCosts ?? 0,
+              schoolValue: trust.schoolAgencySupplyTeachingStaffCosts ?? 0,
+              centralValue: trust.centralAgencySupplyTeachingStaffCosts ?? 0,
+            };
+          }) ?? [],
+        tableHeadings,
+      };
+    }, [data, tableHeadings]);
+
+  const elementId = "teaching-and-teaching-support-staff";
   const [hash] = useHash();
 
   return (
@@ -155,58 +163,62 @@ export const NonEducationalSupportStaff: React.FC<{
           <h2 className="govuk-accordion__section-heading">
             <span
               className="govuk-accordion__section-button"
-              id="accordion-heading-2"
+              id="accordion-heading-1"
             >
-              Non-educational support staff
+              Teaching and teaching support staff
             </span>
           </h2>
         </div>
         <div
-          id="accordion-content-2"
+          id="accordion-content-1"
           className="govuk-accordion__section-content"
-          aria-labelledby="accordion-2"
+          aria-labelledby="accordion-heading-1"
           role="region"
         >
           <HorizontalBarChartWrapper
-            data={totalNonEducationalBarData}
-            chartName="total non-educational support staff costs"
+            data={totalTeachingBarData}
+            chartName="total teaching and support staff cost"
           >
             <h3 className="govuk-heading-s">
-              Total non-educational support staff costs
+              Total teaching and teaching support staff costs
             </h3>
             <ChartDimensions
               dimensions={CostCategories}
               handleChange={handleSelectChange}
-              elementId="total-non-educational-support-staff-costs"
+              elementId="total-teaching-support-staff-cost"
               defaultValue={dimension.value}
             />
           </HorizontalBarChartWrapper>
           <HorizontalBarChartWrapper
-            data={administrativeClericalBarData}
-            chartName="administrative and clerical staff costs"
+            data={teachingStaffBarData}
+            chartName="teaching staff costs"
+          >
+            <h3 className="govuk-heading-s">Teaching staff costs</h3>
+          </HorizontalBarChartWrapper>
+          <HorizontalBarChartWrapper
+            data={supplyTeachingBarData}
+            chartName="supply teaching staff costs"
+          >
+            <h3 className="govuk-heading-s">Supply teaching staff costs</h3>
+          </HorizontalBarChartWrapper>
+          <HorizontalBarChartWrapper
+            data={educationalConsultancyBarData}
+            chartName="educational consultancy costs"
+          >
+            <h3 className="govuk-heading-s">Educational consultancy costs</h3>
+          </HorizontalBarChartWrapper>
+          <HorizontalBarChartWrapper
+            data={educationSupportStaffBarData}
+            chartName="educational support staff costs"
+          >
+            <h3 className="govuk-heading-s">Educational support staff costs</h3>
+          </HorizontalBarChartWrapper>
+          <HorizontalBarChartWrapper
+            data={agencySupplyBarData}
+            chartName="agency supply teaching staff costs"
           >
             <h3 className="govuk-heading-s">
-              Administrative and clerical staff costs
-            </h3>
-          </HorizontalBarChartWrapper>
-          <HorizontalBarChartWrapper
-            data={auditorsCostsBarData}
-            chartName="auditors costs"
-          >
-            <h3 className="govuk-heading-s">Auditors costs</h3>
-          </HorizontalBarChartWrapper>
-          <HorizontalBarChartWrapper
-            data={otherStaffCostsBarData}
-            chartName="other staff costs"
-          >
-            <h3 className="govuk-heading-s">Other staff costs</h3>
-          </HorizontalBarChartWrapper>
-          <HorizontalBarChartWrapper
-            data={professionalServicesBarData}
-            chartName="profession services (non-curriculum) costs"
-          >
-            <h3 className="govuk-heading-s">
-              Professional services (non-curriculum) costs
+              Agency supply teaching staff costs
             </h3>
           </HorizontalBarChartWrapper>
         </div>
