+++ front-end-components/src/views/compare-your-trust/partials/accordion-sections/types.tsx
