+++ front-end-components/src/views/compare-your-trust/partials/index.tsx
@@ -1,0 +1,5 @@
+/* eslint-disable react-refresh/only-export-components */
+export * from "src/views/compare-your-trust/partials/total-expenditure";
+export * from "src/views/compare-your-trust/partials/expenditure-accordion";
+export * from "src/views/compare-your-trust/partials/spending-section";
+export * from "src/views/compare-your-trust/partials/types";
