+++ front-end-components/src/views/compare-your-trust/partials/spending-section.tsx
@@ -2,30 +2,24 @@ import React, { useState } from "react";
 import {
   TotalExpenditure,
   ExpenditureAccordion,
-} from "src/views/compare-your-costs-trust/partials";
-import { CompareYourCostsTrustViewProps } from "src/views/compare-your-costs-trust";
+  SpendingSectionProps,
+} from "src/views/compare-your-trust/partials";
 import { ChartMode, ChartModeChart } from "src/components";
 import {
   SelectedEstablishmentContext,
   ChartModeContext,
   IncludeBreakdownContext,
 } from "src/contexts";
-import { useGovUk } from "src/hooks/useGovUk";
 import {
   BreakdownInclude,
   IncludeBreakdown,
 } from "src/components/include-breakdown";
 
-export const CompareYourCostsTrust: React.FC<CompareYourCostsTrustViewProps> = (
-  props
-) => {
-  const { type, id } = props;
+export const SpendingSection: React.FC<SpendingSectionProps> = ({ id }) => {
   const [displayMode, setDisplayMode] = useState<string>(ChartModeChart);
   const [breakdown, setBreakdown] = useState<string | undefined>(
     BreakdownInclude
   );
-
-  useGovUk();
 
   const toggleChartMode = (e: React.ChangeEvent<HTMLInputElement>) => {
     setDisplayMode(e.target.value);
@@ -50,8 +44,8 @@ export const CompareYourCostsTrust: React.FC<CompareYourCostsTrustViewProps> = (
       </div>
       <ChartModeContext.Provider value={displayMode}>
         <IncludeBreakdownContext.Provider value={breakdown}>
-          <TotalExpenditure id={id} type={type} />
-          <ExpenditureAccordion id={id} type={type} />
+          <TotalExpenditure id={id} />
+          <ExpenditureAccordion id={id} />
         </IncludeBreakdownContext.Provider>
       </ChartModeContext.Provider>
     </SelectedEstablishmentContext.Provider>
