+++ front-end-components/src/views/compare-your-trust/partials/total-expenditure.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { TotalExpenditureData } from "src/views/compare-your-costs-trust/partials";
+import { TotalExpenditureData } from "src/views/compare-your-trust/partials";
 import { ChartDimensionContext } from "src/contexts";
 import {
   CostCategories,
@@ -14,21 +14,19 @@ import {
 import { TrustExpenditure, ExpenditureApi } from "src/services";
 
 export const TotalExpenditure: React.FC<{
-  type: string;
   id: string;
-}> = ({ type, id }) => {
+}> = ({ id }) => {
   const [dimension, setDimension] = useState(PoundsPerPupil);
   const [data, setData] = useState<TrustExpenditure[] | null>();
   const getData = useCallback(async () => {
     setData(null);
     return await ExpenditureApi.trust(
-      type,
       id,
       dimension.value,
       "TotalExpenditure",
       true
     );
-  }, [id, dimension, type]);
+  }, [id, dimension]);
 
   useEffect(() => {
     getData().then((result) => {
