+++ front-end-components/src/views/compare-your-trust/types.tsx
@@ -1,0 +1,3 @@
+export type CompareYourTrustViewProps = {
+  id: string;
+};
