+++ front-end-components/src/views/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react-refresh/only-export-components */
 export * from "src/views/compare-your-costs";
-export * from "src/views/compare-your-costs-trust";
+export * from "src/views/compare-your-trust";
 export * from "src/views/compare-your-census";
 export * from "src/views/historic-data";
 export * from "src/views/deployment-plan";
